organize workflow folders
