+++ packages/core/core-flows/src/order/workflows/exchange/exchange-request-item-return.ts
@@ -1,7 +1,7 @@
 import {
   OrderChangeDTO,
-  OrderClaimDTO,
   OrderDTO,
+  OrderExchangeDTO,
   OrderWorkflow,
   ReturnDTO,
 } from "@medusajs/types"
@@ -13,81 +13,82 @@ import {
   transform,
   when,
 } from "@medusajs/workflows-sdk"
-import { useRemoteQueryStep } from "../../common"
-import { createOrderChangeActionsStep } from "../steps/create-order-change-actions"
-import { createReturnsStep } from "../steps/create-returns"
-import { previewOrderChangeStep } from "../steps/preview-order-change"
-import { updateOrderClaimsStep } from "../steps/update-order-claims"
+import { useRemoteQueryStep } from "../../../common"
+import { createOrderChangeActionsStep } from "../../steps/create-order-change-actions"
+import { createReturnsStep } from "../../steps/create-returns"
+import { previewOrderChangeStep } from "../../steps/preview-order-change"
+import { updateOrderExchangesStep } from "../../steps/update-order-exchanges"
 import {
   throwIfIsCancelled,
   throwIfItemsDoesNotExistsInOrder,
   throwIfOrderChangeIsNotActive,
   throwIfOrderIsCancelled,
-} from "../utils/order-validation"
+} from "../../utils/order-validation"
 
 const validationStep = createStep(
-  "claim-request-item-return-validation",
+  "exchange-request-item-return-validation",
   async function ({
     order,
     orderChange,
     orderReturn,
-    orderClaim,
+    orderExchange,
     items,
   }: {
     order: OrderDTO
     orderReturn: ReturnDTO
-    orderClaim: OrderClaimDTO
+    orderExchange: OrderExchangeDTO
     orderChange: OrderChangeDTO
-    items: OrderWorkflow.OrderClaimRequestItemReturnWorkflowInput["items"]
+    items: OrderWorkflow.OrderExchangeRequestItemReturnWorkflowInput["items"]
   }) {
     throwIfOrderIsCancelled({ order })
-    throwIfIsCancelled(orderClaim, "Claim")
+    throwIfIsCancelled(orderExchange, "Exchange")
     throwIfIsCancelled(orderReturn, "Return")
     throwIfOrderChangeIsNotActive({ orderChange })
     throwIfItemsDoesNotExistsInOrder({ order, inputItems: items })
   }
 )
 
-export const orderClaimRequestItemReturnWorkflowId = "claim-request-item-return"
-export const orderClaimRequestItemReturnWorkflow = createWorkflow(
-  orderClaimRequestItemReturnWorkflowId,
+export const orderExchangeRequestItemReturnWorkflowId =
+  "exchange-request-item-return"
+export const orderExchangeRequestItemReturnWorkflow = createWorkflow(
+  orderExchangeRequestItemReturnWorkflowId,
   function (
-    input: WorkflowData<OrderWorkflow.OrderClaimRequestItemReturnWorkflowInput>
+    input: WorkflowData<OrderWorkflow.OrderExchangeRequestItemReturnWorkflowInput>
   ): WorkflowData<OrderDTO> {
-    const orderClaim = useRemoteQueryStep({
-      entry_point: "order_claim",
+    const orderExchange = useRemoteQueryStep({
+      entry_point: "order_exchange",
       fields: ["id", "order_id", "return_id"],
-      variables: { id: input.claim_id },
+      variables: { id: input.exchange_id },
       list: false,
       throw_if_key_not_found: true,
-    }).config({ name: "claim-query" })
+    }).config({ name: "exchange-query" })
 
-    const existingOrderReturn = when({ orderClaim }, ({ orderClaim }) => {
-      return orderClaim.return_id
+    const existingOrderReturn = when({ orderExchange }, ({ orderExchange }) => {
+      return orderExchange.return_id
     }).then(() => {
       return useRemoteQueryStep({
         entry_point: "return",
         fields: ["id", "status", "order_id"],
-        variables: { id: orderClaim.return_id },
+        variables: { id: orderExchange.return_id },
         list: false,
         throw_if_key_not_found: true,
       }).config({ name: "return-query" }) as ReturnDTO
     })
 
-    const createdReturn = when({ orderClaim }, ({ orderClaim }) => {
-      return !orderClaim.return_id
+    const createdReturn = when({ orderExchange }, ({ orderExchange }) => {
+      return !orderExchange.return_id
     }).then(() => {
       return createReturnsStep([
         {
-          order_id: orderClaim.order_id,
-          claim_id: orderClaim.id,
+          order_id: orderExchange.order_id,
+          exchange_id: orderExchange.id,
         },
       ])
     })
 
     const orderReturn: ReturnDTO = transform(
-      { createdReturn, existingOrderReturn, orderClaim },
-      ({ createdReturn, existingOrderReturn, orderClaim }) => {
+      { createdReturn, existingOrderReturn, orderExchange },
+      ({ createdReturn, existingOrderReturn, orderExchange }) => {
         return existingOrderReturn ?? (createdReturn[0] as ReturnDTO)
       }
     )
@@ -95,7 +96,7 @@ export const orderClaimRequestItemReturnWorkflow = createWorkflow(
     const order: OrderDTO = useRemoteQueryStep({
       entry_point: "orders",
       fields: ["id", "status", "items.*"],
-      variables: { id: orderClaim.order_id },
+      variables: { id: orderExchange.order_id },
       list: false,
       throw_if_key_not_found: true,
     }).config({ name: "order-query" })
@@ -103,20 +104,20 @@ export const orderClaimRequestItemReturnWorkflow = createWorkflow(
     const orderChange: OrderChangeDTO = useRemoteQueryStep({
       entry_point: "order_change",
       fields: ["id", "status"],
-      variables: { order_id: orderClaim.order_id },
+      variables: { order_id: orderExchange.order_id },
       list: false,
     }).config({ name: "order-change-query" })
 
     validationStep({
       order,
       items: input.items,
-      orderClaim,
+      orderExchange,
       orderReturn,
       orderChange,
     })
 
-    when({ orderClaim }, ({ orderClaim }) => {
-      return !orderClaim.return_id
+    when({ orderExchange }, ({ orderExchange }) => {
+      return !orderExchange.return_id
     }).then(() => {
       const createdReturnId = transform(
         { createdReturn },
@@ -124,21 +125,21 @@ export const orderClaimRequestItemReturnWorkflow = createWorkflow(
           return createdReturn[0]!.id
         }
       )
-      updateOrderClaimsStep([
+      updateOrderExchangesStep([
         {
-          id: orderClaim.id,
+          id: orderExchange.id,
           return_id: createdReturnId,
         },
       ])
     })
 
     const orderChangeActionInput = transform(
-      { order, orderChange, orderClaim, orderReturn, items: input.items },
-      ({ order, orderChange, orderClaim, orderReturn, items }) => {
+      { order, orderChange, orderExchange, orderReturn, items: input.items },
+      ({ order, orderChange, orderExchange, orderReturn, items }) => {
         return items.map((item) => ({
           order_change_id: orderChange.id,
           order_id: order.id,
-          claim_id: orderClaim.id,
+          exchange_id: orderExchange.id,
           return_id: orderReturn.id,
           version: orderChange.version,
           action: ChangeActionType.RETURN_ITEM,
@@ -156,6 +157,6 @@ export const orderClaimRequestItemReturnWorkflow = createWorkflow(
 
     createOrderChangeActionsStep(orderChangeActionInput)
 
-    return previewOrderChangeStep(orderClaim.order_id)
+    return previewOrderChangeStep(orderExchange.order_id)
   }
 )
