+++ packages/core/core-flows/src/order/workflows/index.ts
@@ -1,35 +1,35 @@
 export * from "./add-line-items"
 export * from "./archive-orders"
-export * from "./begin-order-claim"
-export * from "./begin-order-exchange"
-export * from "./begin-return"
 export * from "./cancel-order"
 export * from "./cancel-order-change"
 export * from "./cancel-order-fulfillment"
-export * from "./cancel-return"
-export * from "./claim-request-item-return"
+export * from "./claim/begin-order-claim"
+export * from "./claim/claim-request-item-return"
 export * from "./complete-orders"
-export * from "./confirm-return-request"
 export * from "./create-complete-return"
-export * from "./create-exchange-return-shipping-method"
 export * from "./create-fulfillment"
 export * from "./create-order-change"
 export * from "./create-order-change-actions"
 export * from "./create-orders"
-export * from "./create-return-shipping-method"
 export * from "./create-shipment"
 export * from "./decline-order-change"
 export * from "./delete-order-change"
 export * from "./delete-order-change-actions"
-export * from "./exchange-add-new-item"
-export * from "./exchange-request-item-return"
+export * from "./exchange/begin-order-exchange"
+export * from "./exchange/create-exchange-return-shipping-method"
+export * from "./exchange/exchange-add-new-item"
+export * from "./exchange/exchange-request-item-return"
 export * from "./get-order-detail"
 export * from "./get-orders-list"
-export * from "./receive-return"
-export * from "./remove-request-item-return"
-export * from "./remove-return-shipping-method"
-export * from "./request-item-return"
+export * from "./return/begin-return"
+export * from "./return/cancel-return"
+export * from "./return/confirm-return-request"
+export * from "./return/create-return-shipping-method"
+export * from "./return/receive-return"
+export * from "./return/remove-request-item-return"
+export * from "./return/remove-return-shipping-method"
+export * from "./return/request-item-return"
+export * from "./return/update-request-item-return"
+export * from "./return/update-return-shipping-method"
 export * from "./update-order-change-actions"
-export * from "./update-request-item-return"
-export * from "./update-return-shipping-method"
 export * from "./update-tax-lines"
