+++ packages/core/core-flows/src/order/workflows/return/begin-return.ts
@@ -5,10 +5,10 @@ import {
   createWorkflow,
   transform,
 } from "@medusajs/workflows-sdk"
-import { useRemoteQueryStep } from "../../common"
-import { createOrderChangeStep } from "../steps/create-order-change"
-import { createReturnsStep } from "../steps/create-returns"
-import { throwIfOrderIsCancelled } from "../utils/order-validation"
+import { useRemoteQueryStep } from "../../../common"
+import { createOrderChangeStep } from "../../steps/create-order-change"
+import { createReturnsStep } from "../../steps/create-returns"
+import { throwIfOrderIsCancelled } from "../../utils/order-validation"
 
 const validationStep = createStep(
   "begin-return-order-validation",
